financial_shock_engine/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
pytest>=7.4.0
